Remove bigint as id.

+++ satim-app/server/src/db/migrations/001/create_projects_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE projects (
-                          id           BIGSERIAL PRIMARY KEY,
+                          id           SERIAL PRIMARY KEY,
                           name         VARCHAR(40)    NOT NULL,
                           description  VARCHAR(255)             NULL,
-                          owner_id     BIGINT           NOT NULL REFERENCES users(id),
+                          owner_id     INT           NOT NULL REFERENCES users(id),
                           created_at   TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now(),
                           updated_at   TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now()
 );
